[CHORE] my page view model test setup (#199)
